updated tests to vs 2015 and fix warning

+++ tests/utils.cpp
@@ -41,5 +41,5 @@ std::size_t get_filesize(const std::string& filename)
     if (!ifs)
         throw std::runtime_error("Failed open file ");
 
-    return ifs.tellg();
+    return static_cast<std::size_t>(ifs.tellg());
 }
